refactored storre
